requirements.txt
discord.py-self==1.9.2
requests==2.31.0